Change the namespace of the Converter classes
