+++ spec/converters/spree/retail/shopify/image_converter_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-module Shopify
+module Spree::Retail::Shopify
   RSpec.describe ImageConverter do
     include_context 'spree_builders'
